requirements.txt
Flask==1.1.2
Flask-SocketIO==5.0.1
gunicorn==20.0.4
python-dotenv==0.10.1
radar-python==0.0.4
gevent-websocket==0.10.1
Flask-Cors==3.0.9